requirements.txt
# Requirements
tweepy
dotenv
numpy
pillow